+++ DfE.FindInformationAcademiesTrusts/Pages/Trusts/Governance/TrustLeadership.cshtml.cs
@@ -1,0 +1,10 @@
+﻿using DfE.FindInformationAcademiesTrusts.Services.DataSource;
+using DfE.FindInformationAcademiesTrusts.Services.Trust;
+
+namespace DfE.FindInformationAcademiesTrusts.Pages.Trusts.Governance;
+
+public class TrustLeadershipModel(
+    IDataSourceService dataSourceService,
+    ILogger<TrustLeadershipModel> logger,
+    ITrustService trustService)
+    : GovernanceAreaModel(dataSourceService, trustService, logger);